add moar tests
